requirements.txt
pyyaml
msgpack
requests
pymetasploit3
xmltodict
pexpect
bs4
colorama